+++ cobs/settings.hpp
@@ -16,7 +16,7 @@
 namespace cobs {
 
 //! run COBS using parallel threads, default: all cores
-extern uint64_t gopt_threads;
+extern unsigned gopt_threads;
 
 //! whether to load the complete index to RAM for queries.
 extern bool gopt_load_complete_index;
